requirement.txt
PyYAML==6.0
soundfile==0.11.0
torch==1.13.1
pesq==0.0.4
pystoi==0.3.3
einops==0.6.0
scipy==1.10.0
asteroid==0.6.0